deprecate save as draft on initial proposal
