+++ src/app/[orgSlug]/study/[studyId]/proposal/schema.ts
@@ -1,5 +1,4 @@
 import { z } from 'zod'
-import { type UseFormReturnType } from '@mantine/form'
 import { extractTextFromLexical, countWordsFromLexical } from '@/lib/lexical'
 
 const WORD_LIMIT_ERROR = 'Word limit exceeded. Please shorten your text.'
@@ -74,16 +73,11 @@ export const proposalFormSchema = z.object({
 
 export type ProposalFormValues = z.infer<typeof proposalFormSchema>
 
-// The non-lexical proposal fields. These are synced individually through the Yjs
-// fields map (see useYjsFormMap), and are the only fields tracked for the "Save as
-// draft" dirty state: the lexical editor fields auto-save to Yjs continuously
+// The non-lexical proposal fields, synced individually through the Yjs fields map
+// (see useYjsFormMap). The lexical editor fields auto-save to Yjs continuously.
 export const COLLAB_FIELD_KEYS = ['title', 'datasets', 'piName', 'piUserId'] as const
 
 export type CollabFieldKey = (typeof COLLAB_FIELD_KEYS)[number]
-
-export function isProposalDraftDirty(form: UseFormReturnType<ProposalFormValues>): boolean {
-    return COLLAB_FIELD_KEYS.some((field) => form.isDirty(field))
-}
 
 export const initialProposalValues: ProposalFormValues = {
     title: '',
